+++ src/main/java/com/github/pagehelper/sqlsource/OrderByStaticSqlSource.java
@@ -1,7 +1,7 @@
-package com.github.orderbyhelper.sqlsource;
+package com.github.pagehelper.sqlsource;
 
-import com.github.orderbyhelper.OrderByParser;
 import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.parser.OrderByParser;
 import org.apache.ibatis.builder.StaticSqlSource;
 import org.apache.ibatis.mapping.BoundSql;
 import org.apache.ibatis.mapping.ParameterMapping;
@@ -18,12 +18,13 @@ import java.util.List;
  * @author liuzh
  * @since 2015-06-26
  */
-public class OrderByStaticSqlSource implements SqlSource, OrderBySqlSource {
+public class OrderByStaticSqlSource implements SqlSource {
     private String sql;
     private List<ParameterMapping> parameterMappings;
     private Configuration configuration;
     private SqlSource original;
 
+    @SuppressWarnings("unchecked")
     public OrderByStaticSqlSource(StaticSqlSource sqlSource) {
         MetaObject metaObject = SystemMetaObject.forObject(sqlSource);
         this.sql = (String) metaObject.getValue("sql");