+++ src/main/java/com/github/pagehelper/sqlsource/PageSqlSource.java
@@ -1,6 +1,7 @@
 package com.github.pagehelper.sqlsource;
 
 import com.github.pagehelper.SqlUtil;
+import com.github.pagehelper.parser.Parser;
 import org.apache.ibatis.mapping.BoundSql;
 import org.apache.ibatis.mapping.SqlSource;
 
@@ -11,6 +12,12 @@ import org.apache.ibatis.mapping.SqlSource;
  * @since 2015-06-29
  */
 public abstract class PageSqlSource implements SqlSource {
+
+    protected static final ThreadLocal<Parser> localParser = new ThreadLocal<Parser>();
+
+    public void setParser(Parser parser) {
+        localParser.set(parser);
+    }
 
     /**
      * 返回值null - 普通,true - count,false - page
@@ -27,7 +34,7 @@ public abstract class PageSqlSource implements SqlSource {
      * @param parameterObject
      * @return
      */
-    protected abstract BoundSql getDefaultBoundSql(Object  parameterObject);
+    protected abstract BoundSql getDefaultBoundSql(Object parameterObject);
 
     /**
      * 获取Count查询的BoundSql
@@ -35,7 +42,7 @@ public abstract class PageSqlSource implements SqlSource {
      * @param parameterObject
      * @return
      */
-    protected abstract BoundSql getCountBoundSql(Object  parameterObject);
+    protected abstract BoundSql getCountBoundSql(Object parameterObject);
 
     /**
      * 获取分页查询的BoundSql
@@ -43,7 +50,7 @@ public abstract class PageSqlSource implements SqlSource {
      * @param parameterObject
      * @return
      */
-    protected abstract BoundSql getPageBoundSql(Object  parameterObject);
+    protected abstract BoundSql getPageBoundSql(Object parameterObject);
 
     /**
      * 获取BoundSql
@@ -54,9 +61,9 @@ public abstract class PageSqlSource implements SqlSource {
     @Override
     public BoundSql getBoundSql(Object parameterObject) {
         Boolean count = getCount();
-        if(count == null){
+        if (count == null) {
             return getDefaultBoundSql(parameterObject);
-        } else if(count){
+        } else if (count) {
             return getCountBoundSql(parameterObject);
         } else {
             return getPageBoundSql(parameterObject);