+++ src/test/java/com/github/pagehelper/test/basic/dynamic/Where.java
@@ -30,7 +30,7 @@ import java.util.Map;
  * Created by liuzh on 2015/1/11.
  */
 public class Where {
-    private Map<String,Object> map;
+    private Map<String, Object> map;
 
     public Where(Map<String, Object> map) {
         this.map = map;